backend/requirements.txt
fastapi==0.111.0
uvicorn==0.21.1
python-multipart==0.0.9
numpy>=1.26.4
rasterio==1.4.1
shapely==2.0.4
sentinelhub==3.11.3
python-dotenv==1.0.0
pydantic>=2.5.3
pydantic-settings>=2.2.1
pillow==9.5.0
pyproj==3.6.1
psycopg2-binary==2.9.9
